(#268) moving files to other places
