--- blueprints/structural_sections/plotter/__init__.py
@@ -1,1 +1,0 @@
-"""Sub-package for plotting structural sections."""
